requirements.txt
Package               Version 
--------------------- --------
amqp                  1.4.9   
anyjson               0.3.3   
billiard              3.3.0.23
celery                3.1.25  
celery-with-redis     3.0     
Django                1.8.2   
django-celery         3.2.2   
django-haystack       2.8.1   
django-redis-sessions 0.5.6   
django-tinymce        2.6.0   
jieba                 0.39    
kombu                 3.0.37  
Pillow                3.4.1   
pip                   10.0.1  
PyMySQL               0.8.1   
pytz                  2018.4  
redis                 2.10.6  
redis-py-cluster      1.3.4   
setuptools            39.2.0  
uWSGI                 2.0.17.1
wheel                 0.31.1  
Whoosh                2.7.4   
